Create M4-response
First, I will point out that you left notes about the errors in your code. However, for the sake of argument, I first noticed and fixed that our bracket and/or brace "{" was in the wrong location and simply needed to be brought down one line to function properly from line 4 to line 5. The second error that caught my attention was the infamous missing semi-colon ";" at the end of line 23. I have been doing some outside research looking for videos that talk about our textbook, and one that I found mentions that missing a semi-colon is one of the most common and easiest mistakes he sees with new students and programmers in C++, so that is an outstanding error to push, in my opinion, because if we constantly are thinking about the most common errors like forgetting a semi-colon or a brace not being on the right line, that means we will notice because of things like this where we are actively searching enough for it to be a habit which will undoubtedly help us in the future.